backend/requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
httpx
openai 
python-dotenv
